requirements.txt
configparser==5.3.0
laiarturs-ros-api==1.0.0
openpyxl==3.1.2
pandas==1.5.3
paramiko==3.1.0
prettytable==3.6.0
pyfiglet==0.8.post1
